changes: images added for champ principal 🤩
