+++ package.json
@@ -15,6 +15,7 @@
     "react-dom": "^18.2.0",
     "react-i18next": "^12.2.0",
     "react-icons": "^4.8.0",
+    "react-router-dom": "^6.11.1",
     "sweetalert2": "^11.7.3"
   },
   "devDependencies": {
