+++ src/models/json/champs.json
@@ -20,11 +20,11 @@
   { "id": 18, "name": "Caitlyn" },
   { "id": 19, "name": "Camille" },
   { "id": 20, "name": "Cassiopeia" },
-  { "id": 21, "name": "Cho’Gath" },
+  { "id": 21, "name": "Chogath" },
   { "id": 22, "name": "Corki" },
   { "id": 23, "name": "Darius" },
   { "id": 24, "name": "Diana" },
-  { "id": 25, "name": "Dr. Mundo" },
+  { "id": 25, "name": "DrMundo" },
   { "id": 26, "name": "Draven" },
   { "id": 27, "name": "Ekko" },
   { "id": 28, "name": "Elise" },
@@ -51,7 +51,7 @@
   { "id": 49, "name": "Jayce" },
   { "id": 50, "name": "Jhin" },
   { "id": 51, "name": "Jinx" },
-  { "id": 52, "name": "Kai’Sa" },
+  { "id": 52, "name": "KaiSa" },
   { "id": 53, "name": "Kalista" },
   { "id": 54, "name": "Karma" },
   { "id": 55, "name": "Karthus" },
@@ -60,25 +60,25 @@
   { "id": 58, "name": "Kayle" },
   { "id": 59, "name": "Kayn" },
   { "id": 60, "name": "Kennen" },
-  { "id": 61, "name": "Kha’Zix" },
+  { "id": 61, "name": "Khazix" },
   { "id": 62, "name": "Kindred" },
   { "id": 63, "name": "Kled" },
-  { "id": 64, "name": "Kog’Maw" },
-  { "id": 65, "name": "K'Sante" },
+  { "id": 64, "name": "KogMaw" },
+  { "id": 65, "name": "KSante" },
   { "id": 66, "name": "LeBlanc" },
-  { "id": 67, "name": "Lee Sin" },
+  { "id": 67, "name": "LeeSin" },
   { "id": 68, "name": "Leona" },
   { "id": 69, "name": "Lillia" },
   { "id": 70, "name": "Lissandra" },
   { "id": 71, "name": "Lucian" },
   { "id": 72, "name": "Lulu" },
   { "id": 73, "name": "Lux" },
-  { "id": 74, "name": "Maestro Yi" },
+  { "id": 74, "name": "MaestroYi" },
   { "id": 75, "name": "Malphite" },
   { "id": 76, "name": "Malzahar" },
   { "id": 77, "name": "Maokai" },
   { "id": 78, "name": "Milio" },
-  { "id": 79, "name": "Miss Fortune" },
+  { "id": 79, "name": "MissFortune" },
   { "id": 80, "name": "Mordekaiser" },
   { "id": 81, "name": "Morgana" },
   { "id": 82, "name": "Nami" },
@@ -99,9 +99,9 @@
   { "id": 97, "name": "Quinn" },
   { "id": 98, "name": "Rakan" },
   { "id": 99, "name": "Rammus" },
-  { "id": 100, "name": "Rek’Sai" },
+  { "id": 100, "name": "RekSai" },
   { "id": 101, "name": "Rell" },
-  { "id": 102, "name": "Renata Glasc" },
+  { "id": 102, "name": "RenataGlasc" },
   { "id": 103, "name": "Renekton" },
   { "id": 104, "name": "Rengar" },
   { "id": 105, "name": "Riven" },
@@ -133,14 +133,14 @@
   { "id": 131, "name": "Tristana" },
   { "id": 132, "name": "Trundle" },
   { "id": 133, "name": "Tryndamere" },
-  { "id": 134, "name": "Twisted Fate" },
+  { "id": 134, "name": "TwistedFate" },
   { "id": 135, "name": "Twitch" },
   { "id": 136, "name": "Udyr" },
   { "id": 137, "name": "Urgot" },
   { "id": 138, "name": "Varus" },
   { "id": 139, "name": "Vayne" },
   { "id": 140, "name": "Veigar" },
-  { "id": 141, "name": "Vel’Koz" },
+  { "id": 141, "name": "Velkoz" },
   { "id": 142, "name": "Vex" },
   { "id": 143, "name": "Vi" },
   { "id": 144, "name": "Viego" },
@@ -151,7 +151,7 @@
   { "id": 149, "name": "Wukong" },
   { "id": 150, "name": "Xayah" },
   { "id": 151, "name": "Xerath" },
-  { "id": 152, "name": "Xin Zhao" },
+  { "id": 152, "name": "XinZhao" },
   { "id": 153, "name": "Yasuo" },
   { "id": 154, "name": "Yone" },
   { "id": 155, "name": "Yorick" },
